refactor: Move UnsubscriptionError to Observable.ts
Moves UnsubscriptionError to Observable.ts, since that's the only place it is used.
