--- packages/rxjs/src/internal/util/UnsubscriptionError.ts
@@ -1,19 +1,0 @@
-/**
- * An error thrown when one or more errors have occurred during the
- * `unsubscribe` of a {@link Subscription}.
- */
-export class UnsubscriptionError extends Error {
-  /**
-   * @deprecated Internal implementation detail. Do not construct error instances.
-   * Cannot be tagged as internal: https://github.com/ReactiveX/rxjs/issues/6269
-   */
-  constructor(public errors: any[]) {
-    super(
-      errors
-        ? `${errors.length} errors occurred during unsubscription:
-${errors.map((err, i) => `${i + 1}) ${err.toString()}`).join('\n  ')}`
-        : ''
-    );
-    this.name = 'UnsubscriptionError';
-  }
-}
